+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- **[NEW]** Add `Card` util component, add `QrCard` component and use `Card` in `TripCard`
 - [...]
 
 # v23.2.0 (20/03/2020)